Split RulesetInputManager out to FrameStabilityContainer
